backend/requirements.txt
django>=5.2.8
django-ckeditor>=6.7.3
django-cors-headers>=4.9.0
django-filter>=25.2
django-imagekit>=6.0.0
django-unfold>=0.71.0
djangorestframework>=3.16.1
pillow>=12.0.0
